LayoutTests/imported/w3c/web-platform-tests/referrer-policy/gen/srcdoc.meta/same-origin/xhr.http-expected.txt

Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/xhr.py?redirection=keep-origin&action=purge&key=2b9ff1c3-dbea-43d6-af9f-42348a5c5d93&path=%2Fmixed-content
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/common/security-features/subresource/xhr.py?redirection=keep-origin&action=purge&key=2b9ff1c3-dbea-43d6-af9f-42348a5c5d93&path=%2Fmixed-content due to access control checks.
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/xhr.py?redirection=no-redirect&action=purge&key=8c9dbc7f-cf24-4f77-a9d5-095ef68888d2&path=%2Fmixed-content
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/common/security-features/subresource/xhr.py?redirection=no-redirect&action=purge&key=8c9dbc7f-cf24-4f77-a9d5-095ef68888d2&path=%2Fmixed-content due to access control checks.
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/xhr.py?redirection=swap-origin&action=purge&key=adf32ba1-4310-4a1c-8136-a26aa499cdb0&path=%2Fmixed-content
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/common/security-features/subresource/xhr.py?redirection=swap-origin&action=purge&key=adf32ba1-4310-4a1c-8136-a26aa499cdb0&path=%2Fmixed-content due to access control checks.
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/xhr.py?redirection=keep-origin&action=purge&key=3afb4bbf-7825-442c-a8b0-c80ad3ea9853&path=%2Fmixed-content
CONSOLE MESSAGE: XMLHttpRequest cannot load https://www1.localhost:9443/common/security-features/subresource/xhr.py?redirection=keep-origin&action=purge&key=3afb4bbf-7825-442c-a8b0-c80ad3ea9853&path=%2Fmixed-content due to access control checks.
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/xhr.py?redirection=no-redirect&action=purge&key=a132511c-c453-47bd-82d8-9a4e9e8f5b74&path=%2Fmixed-content
CONSOLE MESSAGE: XMLHttpRequest cannot load https://www1.localhost:9443/common/security-features/subresource/xhr.py?redirection=no-redirect&action=purge&key=a132511c-c453-47bd-82d8-9a4e9e8f5b74&path=%2Fmixed-content due to access control checks.
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/xhr.py?redirection=swap-origin&action=purge&key=61e328b8-57b3-4efc-971f-b1aab23cd6e8&path=%2Fmixed-content
CONSOLE MESSAGE: XMLHttpRequest cannot load https://www1.localhost:9443/common/security-features/subresource/xhr.py?redirection=swap-origin&action=purge&key=61e328b8-57b3-4efc-971f-b1aab23cd6e8&path=%2Fmixed-content due to access control checks.
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/xhr.py?action=purge&key=a72904e0-baf6-4d86-8bc4-274644e0a458&path=%2Fmixed-content
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/xhr.py?action=purge&key=1a0c874e-14bd-47fd-8ff3-6755610bb549&path=%2Fmixed-content

FAIL Referrer Policy: Expects omitted for xhr to cross-http origin and keep-origin redirection from http context. promise_test: Unhandled rejection with value: "Network Error"
FAIL Referrer Policy: Expects omitted for xhr to cross-http origin and no-redirect redirection from http context. promise_test: Unhandled rejection with value: "Network Error"
FAIL Referrer Policy: Expects omitted for xhr to cross-http origin and swap-origin redirection from http context. promise_test: Unhandled rejection with value: "Network Error"
FAIL Referrer Policy: Expects omitted for xhr to cross-https origin and keep-origin redirection from http context. promise_test: Unhandled rejection with value: "Network Error"
FAIL Referrer Policy: Expects omitted for xhr to cross-https origin and no-redirect redirection from http context. promise_test: Unhandled rejection with value: "Network Error"
FAIL Referrer Policy: Expects omitted for xhr to cross-https origin and swap-origin redirection from http context. promise_test: Unhandled rejection with value: "Network Error"
FAIL Referrer Policy: Expects omitted for xhr to same-http origin and swap-origin redirection from http context. promise_test: Unhandled rejection with value: "Network Error"
FAIL Referrer Policy: Expects omitted for xhr to same-https origin and swap-origin redirection from http context. promise_test: Unhandled rejection with value: "Network Error"
PASS Referrer Policy: Expects stripped-referrer for xhr to same-http origin and keep-origin redirection from http context.
PASS Referrer Policy: Expects stripped-referrer for xhr to same-http origin and no-redirect redirection from http context.
